Fix 64-bit pointer truncation crashes and add Android ARM64 support
- Fix 64-bit pointer truncation in HSD_DevCom callbacks, lbMemory heap defragmenter, ARQ requests, and THP frame buffers.
- Align audio buffers to 32-byte boundaries in HSD synth.
- Fix disc pointer (DP) macros for 64-bit systems in lbspdisplay and toy.
- Add Android build configuration supporting 16KB max page size for Android 15+.
- Add GCC launcher script to compile decompilation code with CP932 under bionic.
- Route Android logcat and OSPanic messages, and support disc launching via intent.

+++ extern/aurora/include/dolphin/ar.h
@@ -12,11 +12,11 @@ typedef void (*ARQCallback)(struct ARQRequest* request);
 
 struct ARQRequest {
     /* 0x00 */ struct ARQRequest *next;
-    /* 0x04 */ u32 owner;
+    /* 0x04 */ uintptr_t owner;
     /* 0x08 */ u32 type;
     /* 0x0C */ u32 priority;
-    /* 0x10 */ u32 source;
-    /* 0x14 */ u32 dest;
+    /* 0x10 */ uintptr_t source;
+    /* 0x14 */ uintptr_t dest;
     /* 0x18 */ u32 length;
     /* 0x1C */ ARQCallback callback;
 };
@@ -62,9 +62,9 @@ void* ARGetStorageAddress();
 // ARQ
 void ARQInit(void);
 void ARQReset(void);
-void ARQPostRequest(ARQRequest* request, u32 owner, u32 type, u32 priority, uintptr_t source, uintptr_t dest, u32 length, ARQCallback callback);
+void ARQPostRequest(ARQRequest* request, uintptr_t owner, u32 type, u32 priority, uintptr_t source, uintptr_t dest, u32 length, ARQCallback callback);
 void ARQRemoveRequest(ARQRequest* request);
-void ARQRemoveOwnerRequest(u32 owner);
+void ARQRemoveOwnerRequest(uintptr_t owner);
 void ARQFlushQueue(void);
 void ARQSetChunkSize(u32 size);
 u32 ARQGetChunkSize(void);
